updating cache and helpers
