+++ src/Helpers/StoreBlocksTrait.php
@@ -20,224 +20,11 @@ use EightshiftLibs\Exception\InvalidManifest;
 trait StoreBlocksTrait
 {
 	/**
-	 * Cache.
-	 *
-	 * @var array<mixed>
-	 */
-	private static $cache = [];
-
-	/**
-	 * Cache builder.
-	 *
-	 * @var array<mixed>
-	 */
-	private static $cacheBuilder = [];
-
-	/**
-	 * Cache name.
-	 *
-	 * @var string
-	 */
-	private static $cacheName = '';
-
-	/**
-	 * Cache transient prefix.
-	 *
-	 * @var string
-	 */
-	private static $cacheTransientPrefix = '';
-
-	/**
-	 * Cache version.
-	 *
-	 * @var string
-	 */
-	private static $version = '';
-
-	/**
 	 * Styles key
 	 *
 	 * @var array<mixed>
 	 */
 	public static $styles = [];
-
-	// -----------------------------------------------------
-	// CACHE
-	// -----------------------------------------------------
-
-	/**
-	 * Set cache details.
-	 *
-	 * @param string $cacheTransientPrefix Cache transient prefix.
-	 * @param array<mixed> $cacheBuilder Cache builder.
-	 * @param string $cacheName Cache name.
-	 * @param string $version Cache version.
-	 *
-	 * @return void
-	 */
-	public static function setCacheDetails(
-		string $cacheTransientPrefix,
-		array $cacheBuilder,
-		string $cacheName,
-		string $version
-	): void {
-		self::$cacheBuilder = $cacheBuilder;
-		self::$cacheName = $cacheName;
-		self::$cacheTransientPrefix = $cacheTransientPrefix;
-		self::$version = $version;
-	}
-
-	/**
-	 * Set internal cache.
-	 *
-	 * @return void
-	 */
-	public static function setCache(): void
-	{
-		if (self::$cache) {
-			return;
-		}
-
-		$output = [];
-
-		foreach (self::$cacheBuilder as $type => $value) {
-			$data = \get_transient(self::getCacheName($type));
-
-			if (!$data) {
-				continue;
-			}
-
-			if (!Helpers::isJson($data)) {
-				continue;
-			}
-
-			$data = \json_decode($data, true);
-
-			if (!$data) {
-				continue;
-			}
-
-			$output[$type] = $data;
-		}
-
-		self::$cache = $output;
-	}
-
-	/**
-	 * Get cache.
-	 *
-	 * @return array<mixed>
-	 */
-	public static function getCache(): array
-	{
-		return self::$cache;
-	}
-
-	/**
-	 * Get cache name.
-	 *
-	 * @param string $type Type of the cache.
-	 *
-	 * @return string
-	 */
-	public static function getCacheName(string $type): string
-	{
-		return self::$cacheTransientPrefix . "_{$type}";
-	}
-
-	/**
-	 * Check if cache version is valid.
-	 *
-	 * @return bool
-	 */
-	public static function isCacheVersionValid(): bool
-	{
-		return self::getCacheVersion() === self::$version;
-	}
-
-	/**
-	 * Set version cache.
-	 *
-	 * @return void
-	 */
-	public static function setCacheVersion(): void
-	{
-		$name = self::getCacheName(AbstractManifestCache::VERSION_KEY);
-
-		$cache = \get_transient($name);
-
-		if (!$cache) {
-			\set_transient($name, self::$version);
-		}
-	}
-
-	/**
-	 * Get cache version.
-	 *
-	 * @return string
-	 */
-	public static function getCacheVersion(): string
-	{
-		$cache = \get_transient(self::getCacheName(AbstractManifestCache::VERSION_KEY)) ?: ''; // phpcs:ignore WordPress.PHP.DisallowShortTernary.Found
-
-		if (!$cache) {
-			self::setCacheVersion();
-		}
-
-		return $cache;
-	}
-
-	/**
-	 * Unset cache version.
-	 *
-	 * @return void
-	 */
-	public static function deleteCacheVersion(): void
-	{
-		\delete_transient(self::getCacheName(AbstractManifestCache::VERSION_KEY));
-	}
-
-	/**
-	 * Unset all manifest cache.
-	 *
-	 * @return void
-	 */
-	public static function deleteAllCache(): void
-	{
-		$data = \array_keys(self::$cacheBuilder);
-
-		foreach ($data as $cache) {
-			self::deleteCache($cache);
-		}
-
-		self::deleteCacheVersion();
-	}
-
-	/**
-	 * Unset cache item by type.
-	 *
-	 * @param string $cacheType Type of the cache.
-	 *
-	 * @return void
-	 */
-	public static function deleteCache(string $cacheType): void
-	{
-		\delete_transient(self::getCacheName($cacheType));
-	}
-
-	/**
-	 * Check if we should cache the service classes.
-	 *
-	 * @return bool
-	 */
-	public static function shouldCache(): bool
-	{
-		return !(
-			(\defined('WP_ENVIRONMENT_TYPE') &&
-			(\WP_ENVIRONMENT_TYPE === 'development' || \WP_ENVIRONMENT_TYPE === 'local')) ||
-			\defined('WP_CLI')
-		);
-	}
 
 	// -----------------------------------------------------
 	// BLOCKS
@@ -261,7 +48,7 @@ trait StoreBlocksTrait
 		$filterName = 'es_boilerplate_get_blocks';
 
 		if (\has_filter($filterName)) {
-			$output = \apply_filters($filterName, $output, self::$cacheName);
+			$output = \apply_filters($filterName, $output, CacheTrait::getCacheName());
 		}
 
 		return $output;
